+++ Codigo/api/controllers/EspecialidadeController.ts
@@ -31,6 +31,9 @@ class EspecialidadeController {
 
   // URI de exemplo: http://localhost:3000/api/especialidade/1
   public delete: DeleteRequestHandler = async (request, response) => {
+    const especialidade = await this.Service.getById(Number(request.params.id));
+    if (!especialidade)
+      throw new AppError("Especialidade não encontrada", 404);
     await this.Service.delete(Number(request.params.id))
     .then(dado => {
       response.status(204).json({
